// src/type.ts
export interface Movie{
    adult: boolean;
    backdrop_path: string;
    genre_ids: number[];
    id: Id;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: string;
    title: Title;
    video: boolean;
    vote_average: number;
    vote_count: number;

}

export interface MovieR{
    adult: boolean;
    backdrop_path: string;
    genre_ids: string[];
    id: Id;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: string;
    title: Title;
    video: boolean;
    vote_average: number;
    vote_count: number;

}



export type Id=number;

export type Title=string;

export interface Genre{
id:number,

name:string,

}


export interface GenreActive{
 
    active:boolean,
    name: string
    
    }


    export interface MovieView {
        adult:                 boolean;
        backdrop_path:         string;
        belongs_to_collection: BelongsToCollection;
        budget:                number;
        genres:                Genre[];
        homepage:              string;
        id:                    number;
        imdb_id:               string;
        origin_country:        string[];
        original_language:     string;
        original_title:        string;
        overview:              string;
        popularity:            number;
        poster_path:           string;
        production_companies:  ProductionCompany[];
        production_countries:  ProductionCountry[];
        release_date:          Date;
        revenue:               number;
        runtime:               number;
        spoken_languages:      SpokenLanguage[];
        status:                string;
        tagline:               string;
        title:                 string;
        video:                 boolean;
        vote_average:          number;
        vote_count:            number;
    }
    
    export interface BelongsToCollection {
        id:            number;
        name:          string;
        poster_path:   string;
        backdrop_path: string;
    }
    
    export interface Genre {
        id:   number;
        name: string;
    }
    
    export interface ProductionCompany {
        id:             number;
        logo_path:      string;
        name:           string;
        origin_country: string;
    }
    
    export interface ProductionCountry {
        iso_3166_1: string;
        name:       string;
    }
    
    export interface SpokenLanguage {
        english_name: string;
        iso_639_1:    string;
        name:         string;
    }
    